+++ NEWS.md
@@ -1,3 +1,9 @@
+# topografie 0.2.0
+
+* Allow different datasets
+* Added a North_europe dataset
+* Using different database to get geography shapes from
+
 # topografie 0.1.0
 
 * Initial version of this package.